# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Your Auth API",
    "description": "API de autenticacao e identidade para contas proprietarias, projetos, usuarios finais, sessoes, refresh tokens e API keys de projeto.",
    "version": "0.0.1-SNAPSHOT"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Ambiente local"
    }
  ],
...
  "paths": {
    "/projects/{projectId}/users/{userId}": {
      "get": {
        "tags": [
          "Project Users"
        ],
        "summary": "Busca um usuario final pelo id",
        "operationId": "findById",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario encontrado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDTO"
                }
              }
            }
          },
          "401": {
            "description": "Autenticacao obrigatoria.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Conta sem acesso ao projeto.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Projeto ou usuario nao encontrados.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "accessTokenCookie": []
          }
        ]
      },
      "put": {
        "tags": [
          "Project Users"
        ],
        "summary": "Atualiza um usuario final",
        "operationId": "update",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Usuario atualizado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Corpo da requisicao invalido ou erro de validacao.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Autenticacao obrigatoria.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Conta sem acesso ao projeto.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Projeto ou usuario nao encontrados.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "accessTokenCookie": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Project Users"
        ],
        "summary": "Remove um usuario final",
        "operationId": "delete",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Usuario removido."
          },
          "401": {
            "description": "Autenticacao obrigatoria.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Conta sem acesso ao projeto.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Projeto ou usuario nao encontrados.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "accessTokenCookie": []
          }
        ]
      }
    },
...
    "/projects/{id}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Busca um projeto pelo id",
        "operationId": "findById_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Projeto encontrado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDTO"
                }
              }
            }
          },
          "401": {
            "description": "Autenticacao obrigatoria.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Conta sem acesso ao projeto.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Projeto nao encontrado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "accessTokenCookie": []
          }
        ]
      },
      "put": {
        "tags": [
          "Projects"
        ],
        "summary": "Atualiza um projeto",
        "operationId": "update_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Projeto atualizado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Corpo da requisicao invalido ou erro de validacao.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Autenticacao obrigatoria.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Conta sem acesso ao projeto.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Projeto nao encontrado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "accessTokenCookie": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Projects"
        ],
        "summary": "Remove um projeto",
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Projeto removido."
          },
          "401": {
            "description": "Autenticacao obrigatoria.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Conta sem acesso ao projeto.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Projeto nao encontrado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "accessTokenCookie": []
          }
        ]
      }
    },
...
    "/projects/{projectId}/users": {
      "get": {
        "tags": [
          "Project Users"
        ],
        "summary": "Lista usuarios finais de um projeto",
        "operationId": "findAll",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuarios encontrados.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultUserResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Parametros de paginacao invalidos.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Autenticacao obrigatoria.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Conta sem acesso ao projeto.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Projeto nao encontrado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "accessTokenCookie": []
          }
        ]
      },
      "post": {
        "tags": [
          "Project Users"
        ],
        "summary": "Cria um usuario final em um projeto",
        "operationId": "create_1",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Usuario criado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Corpo da requisicao invalido ou erro de validacao.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Autenticacao obrigatoria.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Conta sem acesso ao projeto.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Projeto nao encontrado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Usuario ja existente no projeto.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "accessTokenCookie": []
          }
        ]
      }
    },
...
    "/auth/refresh": {
      "post": {
        "tags": [
          "Account Authentication"
        ],
        "summary": "Renova a sessao web da conta",
        "description": "Usa o cookie refresh_token para gerar novos tokens e redefinir os cookies HTTP-only.",
        "operationId": "refreshToken",
        "parameters": [
          {
            "name": "refresh_token",
            "in": "cookie",
            "description": "Refresh token HTTP-only recebido no login web.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sessao renovada. Novos cookies access-token e refresh_token sao enviados no header Set-Cookie.",
            "headers": {
              "Set-Cookie": {
                "description": "Define novos cookies HTTP-only access-token e refresh_token.",
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountTokensResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Cookie refresh_token ausente.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Refresh token invalido, expirado ou reutilizado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Conta associada ao refresh token nao encontrada.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Falha ao validar, armazenar ou gerar tokens.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "refreshTokenCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UpdateUserDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 320,
            "minLength": 0
          },
          "password": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "phone": {
            "$ref": "#/components/schemas/PhoneDTO"
          }
        },
        "required": [
          "email"
        ]
      },
      "UserResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "DISABLED",
              "BLOCKED"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastLoginAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastPasswordChangedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastFailedLoginAt": {
            "type": "string",
            "format": "date-time"
          },
          "failedLoginAttempts": {
            "type": "integer",
            "format": "int32"
          },
          "lockedUntil": {
            "type": "string",
            "format": "date-time"
          },
          "lastLoginIpAddress": {
            "type": "string"
          },
          "lastLoginUserAgent": {
            "type": "string"
          },
          "phone": {
            "$ref": "#/components/schemas/PhoneDTO"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "error": {
            "type": "string"
          },
          "time": {
            "type": "string",
            "format": "date-time"
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
...
      "UpdateProjectDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 0
          },
          "description": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "SUSPENDED"
            ]
          },
          "environment": {
            "type": "string",
            "enum": [
              "DEVELOPMENT",
              "PRODUCTION"
            ]
          },
          "tokenAudience": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          }
        },
        "required": [
          "environment",
          "name",
          "status",
          "tokenAudience"
        ]
      },
      "ProjectResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "ownerAccountId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "SUSPENDED"
            ]
          },
          "environment": {
            "type": "string",
            "enum": [
              "DEVELOPMENT",
              "PRODUCTION"
            ]
          },
          "tokenAudience": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateUserDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 320,
            "minLength": 0
          },
          "password": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "phone": {
            "$ref": "#/components/schemas/PhoneDTO"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT de conta proprietaria enviado no header Authorization: Bearer <token>.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "projectApiKey": {
        "type": "apiKey",
        "description": "API key de projeto usada por clientes terceiros. Tambem pode ser enviada como Authorization: Bearer ya_sk_...",
        "name": "X-API-Key",
        "in": "header"
      },
      "accessTokenCookie": {
        "type": "apiKey",
        "description": "Cookie HTTP-only usado pelo fluxo web para autenticacao.",
        "name": "access-token",
        "in": "cookie"
      },
      "refreshTokenCookie": {
        "type": "apiKey",
        "description": "Cookie HTTP-only usado pelo fluxo web para renovar a sessao.",
        "name": "refresh_token",
        "in": "cookie"
      }
    }
  }
}